BLE/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
bidict==0.23.1
bleak==0.22.3
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
docutils==0.21.2
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
jinja2==3.1.6
markdown2==2.5.3
MarkupSafe==3.0.2
multidict==6.4.3
nicegui==2.15.0
orjson==3.10.16
propcache==0.3.1
pscript==0.7.7
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-libdispatch==10.3.2
pyserial==3.5
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
